src/requirements.txt
#
# This file is autogenerated by pip-compile with Python 3.9
# by the following command:
#
#    pip-compile requirements.in
#
amqp==5.1.1
    # via kombu
asgiref==3.6.0
    # via django
async-timeout==4.0.2
    # via redis
attrs==22.2.0
    # via jsonschema
bcrypt==4.0.1
    # via -r requirements.in
billiard==3.6.4.0
    # via celery
celery[redis]==5.2.7
    # via -r requirements.in
celery-progress==0.1.3
    # via -r requirements.in
certifi==2022.12.7
    # via
    #   requests
    #   sentry-sdk
cffi==1.15.1
    # via cryptography
charset-normalizer==3.0.1
    # via requests
click==8.1.3
    # via
    #   celery
    #   click-didyoumean
    #   click-plugins
    #   click-repl
click-didyoumean==0.3.0
    # via celery
click-plugins==1.1.1
    # via celery
click-repl==0.2.0
    # via celery
cryptography==39.0.1
    # via pyjwt
defusedxml==0.7.1
    # via python3-openid
django==4.1.7
    # via
    #   -r requirements.in
    #   django-allauth
    #   django-allauth-2fa
    #   django-anymail
    #   django-otp
    #   djangorestframework
    #   drf-spectacular
django-allauth==0.52.0
    # via
    #   -r requirements.in
    #   django-allauth-2fa
django-allauth-2fa==0.10.0
    # via -r requirements.in
django-anymail[mailgun]==9.0
    # via -r requirements.in
django-environ==0.9.0
    # via -r requirements.in
django-otp==1.1.4
    # via django-allauth-2fa
django-waffle==3.0.0
    # via -r requirements.in
djangorestframework==3.14.0
    # via
    #   -r requirements.in
    #   drf-spectacular
djangorestframework-api-key==2.3.0
    # via -r requirements.in
dnspython==2.3.0
    # via -r requirements.in
drf-spectacular==0.25.1
    # via -r requirements.in
idna==3.4
    # via requests
inflection==0.5.1
    # via drf-spectacular
jsonschema==4.17.3
    # via drf-spectacular
kombu==5.2.4
    # via celery
oauthlib==3.2.2
    # via requests-oauthlib
prompt-toolkit==3.0.38
    # via click-repl
psycopg2-binary==2.9.5
    # via -r requirements.in
pycparser==2.21
    # via cffi
pyjwt[crypto]==2.6.0
    # via django-allauth
pypng==0.20220715.0
    # via qrcode
pyrsistent==0.19.3
    # via jsonschema
python3-openid==3.2.0
    # via django-allauth
pytz==2022.7.1
    # via
    #   celery
    #   djangorestframework
pyyaml==6.0
    # via drf-spectacular
qrcode==7.4.2
    # via django-allauth-2fa
redis==4.5.1
    # via celery
requests==2.28.2
    # via
    #   django-allauth
    #   django-anymail
    #   requests-oauthlib
requests-oauthlib==1.3.1
    # via django-allauth
sentry-sdk==1.15.0
    # via -r requirements.in
six==1.16.0
    # via click-repl
sqlparse==0.4.3
    # via django
typing-extensions==4.5.0
    # via qrcode
uritemplate==4.1.1
    # via drf-spectacular
urllib3==1.26.14
    # via
    #   requests
    #   sentry-sdk
vine==5.0.0
    # via
    #   amqp
    #   celery
    #   kombu
wcwidth==0.2.6
    # via prompt-toolkit
jinja2==3.1.2
click==8.1.3
pydantic==1.10.2
selenium==4.8.2